improvements on osc input
