+++ ToposServer/banner.js
@@ -1,0 +1,14 @@
+var pjson = require('./package.json');
+let banner = `
+┏┳┓         ┏┓┏┓┏┓
+ ┃ ┏┓┏┓┏┓┏  ┃┃┗┓┃ 
+ ┻ ┗┛┣┛┗┛┛  ┗┛┗┛┗┛
+     ┛            
+             ${pjson.version}\n`
+function greet() {
+    console.log(banner)
+}
+
+module.exports = {
+    greet: greet
+}
